🐛Always download candidate meta during resolution
This change makes sure that the collection artifact URL, SHA and the
server token are always cached in the collection artifacts manager by
the time the resolution process completes. As a result, it guarantees
that retrieving an artifact in the tmp dir will always know where to
take it from.